render profile page for ostatus follow url
when initiating a follow from a remote mastadon instance, the user gets directed to 'apps/social/ostatus/follow/?uri=' which is currently broken.

this changes it to render the regular profile page for that link, allowing the user to follow the profile.

I'm not sure what the previous behavior was before things got broken

Signed-off-by: Robin Appelman <robin@icewind.nl>