Retrieving id from response instead of this.accountInfo
As this.accountInfo is coming up as undefined, but the response does contain the id of the account, it should work fine in passing the id from response instead of this.accountInfo. this.accountInfo is undefined for localaccount because it is trying to find username@nextcloud.domain user from state which has username@localhost user. 

Signed-off-by: Srijan Saxena <srijan.saxena0@gmail.com>