Update account.js
The method _getActorIdForAccount was not returning the value. Return statement added. Also the state.accountIdMap contains local accounts as username@localhost. But the account passed to it was username@nextcloud.domain even in case of local accounts. The same has been handled by checking the hostname of account to the current hostname. Should work fine for local account.

Signed-off-by: Srijan Saxena <srijan.saxena0@gmail.com>